requirements.txt
albumentations==1.3.0
matplotlib==3.5.1
numpy==1.21.6
Pillow==9.2.0
PyYAML==6.0
scipy==1.7.3
skimage==0.0
SoundFile==0.10.3.post1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
utils==1.0.1
